Omnibus::Reports.pretty_version_map should display a version 'guid'
The content returned by pretty_version_map currently does not display
enough information to truly reconstruct an Omnibus package. A good
example of this issue would be a software definition that has a git
source with a version set to 'master'.

The easiest way to alleviate this issue is add a version guid to the
output in pretty_version_map.  The contents/source of this version guid
differ for each software component depending on the underlying fetcher
implementation:

* GitFetcher: current revision SHA
* NetFetcher (and S3CacheFetcher): MD5 checksum