Omnibus::Software understands version overrides
This adds an additional @override_version property which can be set independently of the standard version (here renamed to @given_version)

The getter functionality of #version has been modified to return the effective version, taking into account any override information if present.

The change was made in this way to minimize the impact to other code and existing omnibus software descriptor files.