Pretty much rewrite top-level Omnibus + Config
This commit consolidates how Omnibus is configured.  Before, it
required both an omnibus.rb file AND additional undocumented
configuration in a Rakefile.

Now, the omnibus.rb file is completely unnecessary, along with much of
the Rakefile config.  Sane defaults are now in place, and most cases
can get by with simply adding

    Omnibus.configure

to the Rakefile.  If defaults need to be altered, just pass a block:

    Omnibus.configure do |config|
      ...
    end

Any places in existing project / software DSL files that refer to old
configuration parameters that were defined directly on the Omnibus
module will have to be change do refer to `Omnibus.config.$PARAMETER`.

Additionally, Rakefile configurations will need to be altered as well.