make project Rakefile optional; don't shell out to rake
The Rakefile is now optional in all Omnibus projects that are built
using the `omnibus` CLI.

The `omnibus build` command now looks for an optional Omnibus
configuration file at the following locations:

* specified by the user using the `-c|--config` CLI option
* `$PWD/omnibus.rb`

If no valid configuration is found the build is run with the default
configuration (loaded by a call to `Omnibus.configure`). This will allow
a user to do things like have a system-wide Omnibus configuration file
which contains

The default project build Rake task is now invoked directly in the CLI
code which means no more shelling out to run Rake!