Add a Vagrant 1.1 build lab to generated projects
This lab consists of a `Berksfile` and `Vagrantifle` dropped into the
root  of every generated project.

The `omnibus` cookbook (and it's dependencies) now live external to this
repository (as they should). The `omnibus` cookbook also depends on
plain, old upstream community cookbooks and will be released to the
community site.