Add support for a software.library_paths attribute
When using omnibus to build packages that are dependent upon other
packages in the same source base (ie. a "common" package that includes
all shared libraries for its dependents) it would be very handy to
provide a mechanism similar to LD_LIBRARY_PATH. In this case, it is
possible to provide a list of library search paths, instead of
explicitly whitelisting individual files/regexs. These search paths
support glob-style syntax in order to make this functionality as
flexible as possible.

All tests pass.