Use git to filter tags when searching cache
The `git tag -l` can be given a pattern such that only tags matching the
pattern are returned. Instead of retrieving a listing of all tags and
filtering in the ruby code, we now ask for the tag we are looking for
and will get back either a single line (hit) or no output (miss).

This will greatly reduce build job output which was streaming the
tag listing multiple times prior to this patch.