Include software config file content in tag digest
Ensure that a change to how an item is built invalidates the cache. This
is important since a build could succeed and be saved into the cache
only to later fail the health check. In almost all cases, the fix for
this will either be a change to the software or its build config.

When computing the cache key (git tag name), we now compute the SHA256
digest of $SOURCE_CONFIG + $DEP_STRING where $SOURCE_CONFIG is the
contents of the software config and $DEP_STRING is $NAME-$VERSION of all
items built before this one delimited with '-'.

Omnibus::Software has a new attr_reader source_config to allow access to
the file contents of the software config.