Warn when incorrectly using `replaces`
This DSL line was introduced only to deal with the renaming of the "chef-full"
artifacts to "chef" back in the dark ages.  It has been copypasta'd around to
other projects to the point where there was a completely inaccurate @todo added
to make replaces default to a broken configuration. Very few projects should be
using replaces and including a replaces line which is the same as the
package_name will break RPM upgrades.

Patch validates that replaces != package_name and yells at the
user to knock if off if they're doing that.