remove libgcc_s from the whitelist
if we include this it hugely violates omnibus and requires users to have
already installed libgcc in order to install omnibus.  instead what we
do is ship libgcc in $install_dir/embedded/lib/libgcc_s.so and packages
that break on this health check need to add 'dependency libgcc' to
their software definitions.