Refactor "omnibus release" into a non-S3-specific thing
The current release command is heavily tied to S3. This commit extracts
the package-specific bits into a new `Package` class, which removes
much of the logic from the S3 releaser. These methods are now
available in future releasers.

This commit renames the CLI command from "omnibus release" to
"omnibus publish". As a team, we agreed that "release" was an incorrect
verb for the action taking place. NOTE: this commit does not include a
deprecation yet. That will be done in a future commit.