[ISSUE-185]
Simplify net_fetcher logic

Because the omnibus cookbook sets up 7-zip on Windows, and 7-zip can handle
decompression of flat archive formats (7z/zip).  On
Windows, it should be safe to defer to 7-zip in those cases.  Also, if someone is downloading .7z
files on a non-Windows platform then one would use 7z and not 7z.exe.

tar is available on all platforms to handle tarball formats.  Compression schemes are usually linked in
to tar, so that another external dependency on gzip, bzip2, etc. do not have to be added.  Added uncompressed
tar to the supported formats.

Also, had to run code through Rubocop to appease Travis.