Cleanup NetFetcher logic and add better Windows support
Because the Omnibus cookbook sets up 7-zip on Windows, we defer to using Z-zip
to unzip those types of archives. If someone is downloading .7zs on a
non-Windows platform, we will use 7z, not 7z.exe to extract those.

"tar" is available on all platforms, so delegate to that binary for handling.
This commit also adds uncompresed tar to the list of supported formats.