Automatically upload checksums if they are present on the Artifact object
It turns out that upload_with_checksum does not actually set the checksum, but
rather compares against the known package and raises an error if it does not
yet exist. API design is definitely not in JFrog's book of best practices.

This commit changes the way we deploy artifacts to be more magical.