Monkey patch Mash to allow "dot" notation of resource accessors
Ohai 7 does not allow users to call attributes by method names. Previously in
Ohai 6, one could do something like:

    Ohai.kernel.machine

In Ohai 7, you must use Ohai as a hash:

    Ohai['kernel']['machine']

This commit restores the functionality presented with the previous version of
Ohai, while adding a deprecation message that you should not use the "dot"
notation in the future.