Standardize on install_dir
Sorry for the confusion folks, but it seems like the move to install_path was
an unwise one. It is impossible to retain backwards compatability using
install_path, and all the other methods in omnibus use the "_dir" suffix,
especially those in the Config object.

This commit is basically a revert of c4c604b, but has some additional fixes
thus making it not a real "revert".