Move windows_safe_path to Builder
This commit changes the way paths are represented in Omnibus. Prior to
this commit, paths would be operating-system-specific. However, with the
introduction of the new Builder DSL methods like +copy+ and +move+, Dir
globbing expects paths to be Rubyesque.

Now, all paths are represented using forward slashes (like the Unixes,
Darwins, and BSDs), and it is the responsibility of the software
writer to wrap command calls that shellout to the system in a call to
+windows_safe_path+.