Fix bucket retrieval for 2.0-stable
Due to a change in Amazon's API, creating a bucket is no longer
idempotent. If a bucket is created when it already exists
an error is now thrown. This broke omnibus cache commands in 2.0-stable.
The code that exists in new omnibus versions avoids this issue,
so the fix here is just to back port and retrofit that code into
the 2.0-stable branch.