Output the contents of the RPM script, not the path
In the move from FPM to custom packagers, we made an incorrect
assumption that the scripts directive in the RPM spec file contained
the _path_ to the script, not the actual contents. This assumption has
proven to be faulty.

This commit changes the RPM packager and spec to output the contents of
the script file, not the path on disk.