Reset target_shasum on fetch to avoid erroneous cache-misses
If a build produces build artifacts in the target_dir (such as beam
files), the initial shasum in a subsequent build will include these,
leading to a git_cache miss even when nothing has changed.

This does introduce a small performance regression as it takes the
shasum of the target_dir on every fetch even for projects that don't
produce build artifacts.