Respect Config.append_timestamp everywhere
Previously, Config.append_timestamp was only respected by the
Omnibus::BuildVersion.semver function. This option is now respected by
the dsl-based version builder as well, ensuring that even when the
user specifies an explicit version with a string, a timestamp is
appended.

This is a behavior change for some users. Users who do not want a
timstamp can alter their configuration.

We've added backwards compatibility for those using the 'semver' style
versions to ensure that we do not append a timestamp to a version
already containing a timestamp.