[bugfix] extra_pkg_file DSL allows directories
Since omnibus 4, the extra_package_file Project DSL wasn't allowing
directories anymore, which is kind of annoying when you're copying a dir
with a whole bunch of configuration files/examples files.

I added a switch to use `copy_file` or `sync` depending on the type
(file or folder) of the source to copy.

It fixes https://github.com/chef/omnibus/issues/464

Obvious fix.