`Metadata#arch` should return a valid Ohai architecture
The change introduced in #535 began returning architecture values of 
`:x86` and `:x64` on Windows. We would prefer the `*.metadata.json` 
files contain valid Ohai-based values of `i386` or `x86_64` as we use 
this data when publishing to backend systems like Artifactory. This 
saves unnecessary translation when fetching packages at install time.