Allow net_fetcher to extract archives under a directory
Clears up a log of fiddly logic on how fetched archives are extracted.
It used to be that net_fetcher would use relative_path as the directory
under which any non-archive files/directories would be downloaded to.
This was indirectly done because the project_dir passed to fetchers
would be the same as project_dir in the software definition which was
defined as "source_dir/relative_dir" or "source_dir/software.name".

Archives were unzipped directly to Config.source_dir in the hopes that
the source archive was well behaved and would extract into a single
subdirectory that the user could select with relative_path.  This is not
the case with a lot of mingw/msys packages which deploy bin, var, lib,
share etc. to the top level source directory.

Henceforth, we are simplifying the behavior by forcing net_fetcher to
always extract files to project_dir as provided to the fetcher.  Any
single file source is also downloaded as a file underneath this
project_dir.  The project_dir in a fetcher is no longer the same as the
project_dir in a software definition.  The software definition as the
notion of a "package_dir" which is an isolated location where a fetcher
my bring down files from a source, and then independently has the notion
of a project_dir which is the directory under which all build commands
are run.  A project_dir is "package_dir/relative_path".  Hence any path
that was previously provided by the user to work underneath a tar-ball
should continue to work.

The operations of clean and fetch were also simplified similar to the
last 2 commits.  fetch will only ensure that the file cache is
up-to-date.  clean will only ensure that project_dir matches the
contents of the source file/archive in the file cache.

One further cleanup was the use of a temporary directory when unzipping
tar-balls on windows.  bsdtar has issues extracting files which are
marked as r/o inside the tarball.  Hence, we use 7zip which doesn't
expand a tar.gz (or tar.whatever) in one shot.  Previously, the
uncompressed tar-ball would be left alongside the fully extracted
contents at the root of Config.source_dir.  Now that we are isolating
expansion, we want to ensure that project_dir doesn't contain any file
that wasn't in the expanded source archive.  Therefore we land the
tar-ball in a temporary directory and delete it after expansion.