Add `maintainer` to the Software DSL
The maintainer field specifies the human maintainer of the software
definition. The responsibilities of the maintainer are likely a policy
decision of the organization or project maintainer the software
definition. This field is currently unused by omnibus itself but could
be used by projects like omnibus-software to help build simple tooling
to help maintain the project.