Use native platform detection for DEB
Basing the package platform against the kernel machine name
has a lot of pitfalls, especially on DEB based systems. Instead
of having lots of code to try and figure out the correct
architecture, this just asks the packaing tools.
On DEB based systems the command is:
  dpkg --print-architecture
The dpkg tool is always available on DEB based systems.
The output is now cached so it will only ever shellout once.