Allow users to generate a manifest for other platforms.
To use this, run:

```
omnibus manifest --os=<blah> --platform-family=<blah> --platform=<blah> --platform-version=<blah>
```

These values are the same as Ohai platform values, and will cause software
definitions that switch on things like `if windows?` to act as if they are
on that platform, allowing you to generate a Windows manifest for your
project from OS/X or Linux.

This feature is intended to allow you to pre-generate manifests for all
platforms you care about.