Don't publish full version manifest to Artifactory
Publishing a package with a large version manifest fails with the following
cryptic error:

```
/opt/languages/ruby/2.1.5/lib/ruby/2.1.0/net/http/generic_request.rb:202:in
`copy_stream': Broken pipe - sendfile (Errno::EPIPE)
```

The error is related to the way in which properties are set via Artifactory's
REST API. The properties are attached to the upload URI as matrix parameters:
https://www.jfrog.com/confluence/display/RTF3X/Artifactory+REST+API#ArtifactoryRESTAPI-SetItemProperties
https://www.w3.org/DesignIssues/MatrixURIs.html

We'll avoid this issue by publishing a subset of the version manifest data
as the `omnibus.dependencies` property. This lighterweight hash will only
contain a list of dependencies and their associated version and license data.