[JEX-162] Prefer server marketing names on Windows
In order to support the creation of `*.appx` packages we will begin 
executing our Windows Omnibus builds on Windows 2012R2 (currently we use
2008R2). Right now builds executed on Windows 2012R2 generate a 
`*.metadata.json` file with a platform_version of `8.1`. This is because 
Windows 8.1 and Windows Server 2012R2 share the same underlying platform 
version (`6.3.9600`). In order to definitively identify the correct 
marketing name you need to make additional Win32 API function calls. We 
would prefer not to incur the overhead required to query the Win32 API 
directly. We also don't need to be that exact in this situation and 
preferring the Windows server marketing names is "good 'nuff"!

Take a look at Chef's `Chef::ReservedNames::Win32::Version` class for 
some of the fun backflips you have to perform in order to correctly 
identify Windows 8 vs Windows 2012 and Windows 8.1 vs Windows 2012R2:
https://github.com/chef/chef/blob/master/lib/chef/win32/version.rb