Include the name of the dependency instead of the name of the project for transitive dependencies.
Currently the text we include for transitive dependencies is like:

This product includes yum 3.11.0
which is a 'chef_berkshelf' dependency of 'chef-server',
and which is available under a 'Apache 2.0' License.
For details, see:
/opt/opscode/LICENSES/chef_berkshelf-yum-3.11.0-LICENSE

whereas it should be like

This product includes yum 3.11.0
which is a 'chef_berkshelf' dependency of 'private-chef-cookbooks',
and which is available under a 'Apache 2.0' License.
For details, see:
/opt/opscode/LICENSES/chef_berkshelf-yum-3.11.0-LICENSE