Check for the transitive dependency licensing errors during post_build.
Without this change when we have a transitive dependency licensing issue, we need to clear he omnibus caches to recover since we cache the build of a software with transitive dependency licensing errors. With this change as long as the :fatal_transitive_dependency_licensing_warnings is set, we will not need a cache cleanup after a transitive dependency licensing error.

Signed-off-by: Serdar Sutay <serdarsutay@gmail.com>