Don't require build user to have a configured Git identity
I got this failure running `omnibus build`:

    The following shell command exited with status 128:

        $ git -c core.autocrlf=false --git-dir=./local/cache/git_cache/opt/codethink-toolchain --work-tree=/opt/codethink-toolchain commit -q -m "Backup of gcc-648708f4a353b91baa1e5cb62cccc3140495d09efa783d22e44e88ef0104f670-1"

    Output:

        (nothing)

    Error:

        *** Please tell me who you are.

    Run

      git config --global user.email "you@example.com"
      git config --global user.name "Your Name"

    to set your account's default identity.
    Omit --global to set the identity only in this repository.

    fatal: unable to auto-detect email address (got 'builduser@oracle-1.(none)')

Running a package builder shouldn't require me to configure a Git identity
for the account that's being used.

This failure comes from the git_cache module. This uses Git purely as
versioned local storage, none of the commits made there are ever pushed
or shared. So the logical fix is to set up a dummy local identity in
the repo it creates.

This should also fix https://github.com/chef/omnibus/issues/512

Signed-off-by: Sam Thursfield <sam.thursfield@codethink.co.uk>