[git_fetcher] Add option to always fetch tags from remote (#41)
Using this option ensures that whatever version of git is used, we
consistently pull the tags in all cases (even when the latest local
commit is up-to-date with the remote but doesn't have a remote tag).

When the option is set, we fetch tags in a separate git command
because until git v1.9 `git fetch` was not a strict subset of
`git fetch --tags` (see http://stackoverflow.com/a/20608181 for more
details). When we get rid of older versions of git, we'll be able
to only use `git fetch --tags`.