Add option to build command to populate S3 cache.
This new build option allows us to automatically upload missing cache
items to S3 at build time. This is especially useful for binary
dependencies which are platform specific as it allows use to leverage
the running/existing build matrix across platforms and architectures to
ensure the artifacts are properly cached in S3.

Resolves internal issue JEX-518.

Signed-off-by: Ryan Hass <rhass@users.noreply.github.com>