Add fallback to BUILD_TIMESTAMP
This patchs adds a fallback to the previous behavior in case
BUILD_TIMESTAMP variable is not set to don't break backwards
compatibility.

The current behaviour is as follow:

Check if BUILD_TIMESTAMP is set and use it if it's valid or ...
check if BUILD_ID is set and use it if it's a valid timestamp

If the timestamp is not valid it fails.

Signed-off-by: itsuugo <antonio.ojea.garcia@gmail.com>