Loosen the pinning on "ohai"
Ohai requirement of ~> 8.0 is incompatible with Chef 13 or newer.

Unpin the requirement a little to allow for Ohai up to 14.x, as for
omnibus's needs the Ohai API should be stable up through Chef/Omnibus 14

As we may now resolve to a dependency that requires ruby >= 2.3, also
bump Omnibus's minimum required ruby >= 2.3 in the gemspec, updates the
README to specify ruby 2.2, and drops the 2.2 testing from appveyor and
travis.

Signed-off-by: Richard Clamp <richardc@unixbeard.net>