Break dmg creation and assets copying into separate steps
We are transitioning our CI to use Veertu Anka for macos builds.
When `anka run` runs the omnibus build it fails when it shells out
the `hdiutil create` command in the `create_writable_dmg` method.
The `hdiutil create` command creates the dmg and copies assets
into it. Most of the `hdiutil create` command works fine but it
fails with `hdiutil: create failed - Resource busy` when it attempts
to eject the disk used to mount the dmg. The easiest solution is
to use `hdiutil create` to only create the dmg and then copy
the assets in a separate step.

Signed-off-by: Jeremiah Snapp <jeremiah@chef.io>