[NetFetcher] Make downloaded file name unique to the software def
The filename of the downloaded file was the basename of the URL. These
files are all written to the root of the omnibus `cache/` directory.
So if multiple software defs had sources with URLs that had the same
basename, they would overwrite one another's source on disk. This leads
to race conditions since `NetFetcher`s download sources concurrently
for perf reasons.

This makes the filename unique to each software def by prepending it
by the software def name.