Update the rpm package signing to work with newer rpm version
Newer versions of rpmsign (rpm 4.13+), no longer prompt for the pass
phrase, but instead provide gpg with access to stdin/out to request it.

Because this process is deferred, you will also often end up with
warning output from rpmbuild before the prompt, so we can no longer
assume the prompt will be the first output.

This issue was discovered while trying to build rpm packages for
OpenSUSE 15 which has the newer rpm.

Signed-off-by: DJ Mountney <david@twkie.net>