[ship_source] Improve source shipping mechanism by not dling twice the sources (#83)
Previously, when source files needed to be shipped alongside a software, we
downloaded the source to build it, and downloaded it again to add it to the
sources/ folder of the final package. This was a waste of network resources,
and additionally it could cause a build to fail (for example if the site hosting
the source became offline during the packaging) when we could avoid it by
reusing the already-downloaded package.

The way of shipping the source has been changed: the Project tells each Software
where to put its sources (if the sources need to be shipped).
The "ship_source true" command must be added to a software definition before
the source command to indicates that the sources must also be shipped.
After all Softwares have been loaded, the sources are taken from this place and
put into the installation directory, in the sources/ sub-directory.