[health_check] Fix health-check on macOS when @loader_path is used (#86)
Some of our libraries have dependencies located with @loader_path
(in particular python libraries in site-packages). They were whitelisted
in the health-check because otool doesn't expand the @loader_path
special variable, resulting in an error when checking if the
dependencies are present.
This fixes this behaviour by manually expanding @loader_path when
encountered.