Allow ohai versions > 17 to be used
The restricton `< 17` was originally added by 2777c5b65 in
https://github.com/chef/omnibus/pull/1013 because `omnibus-toolchain`
was still shipping Ruby 2.6 at that time, and Ohai 17 required Ruby 17.
However, starting with
https://github.com/chef/omnibus-toolchain/pull/153, `omnibus-toolchain`
is on Ruby 2.7+, and hence this restriction can be removed.

To be on the safe side, we are bumping the upper limit to 18 to avoid
any surprises when next major version of Ohai comes out.

Signed-off-by: Balasankar "Balu" C <balasankar@gitlab.com>