Replace BSD tar with GNU tar on Windows
We are using the latest version of msys2 and it comes with GNU tar.

GNU tar requires --force-local when the path to the local compressed file
has a colon in it.

GNU tar treats the "safe" path's backslash separator as an escape
character which causes problems so instead we will use the default
path separator which is the forward slash.

Signed-off-by: Jeremiah Snapp <jeremiah@chef.io>