Add a small quality-of-life change for tests
In the Project and Library specs, we use an `expect` construct to
ensure the build order is consistent. However, when there is a
failure, we see a diff of the internal object representations.
We can map the `:name` property into the construct instead to get
a diff of the component names.

Signed-off-by: Ryan Egesdahl <deriamis@gmail.com>