Catch up to upstream on fixes for DMG creation (#206)
* Fix bug where volume name was not quoted in DMG compression

Signed-off-by: Scott Hain <shain@chef.io>

* Comment why we use zlib on macOS

Help the next person to avoid spending the investigation time.

Signed-off-by: Tim Smith <tsmith@chef.io>

* Break dmg creation and assets copying into separate steps

We are transitioning our CI to use Veertu Anka for macos builds.
When `anka run` runs the omnibus build it fails when it shells out
the `hdiutil create` command in the `create_writable_dmg` method.
The `hdiutil create` command creates the dmg and copies assets
into it. Most of the `hdiutil create` command works fine but it
fails with `hdiutil: create failed - Resource busy` when it attempts
to eject the disk used to mount the dmg. The easiest solution is
to use `hdiutil create` to only create the dmg and then copy
the assets in a separate step.

Signed-off-by: Jeremiah Snapp <jeremiah@chef.io>

* Separate removing writable dmg

Fix expeditor config

This will resolve the hdiutil resource busy failure
during compress_dmg

Signed-off-by: Jaymala Sinha <jsinha@chef.io>

* Enable hdiutil progress tracking options

puppetstrings option enables progress output on hdiutil should help resolve
resource busy issue

Signed-off-by: Jaymala Sinha <jsinha@chef.io>

* Verify the compressed dmg

Signed-off-by: Jaymala Sinha <jsinha@chef.io>

* Check to confirm hdiutil detach succeeds

Signed-off-by: Jaymala Sinha <jsinha@chef.io>

* Unmount dmg before detaching

Signed-off-by: Jaymala Sinha <jsinha@chef.io>

* Create sleep loop in compress_dmg

This helps avoid failure when hdiutil is not
able to successfully detach the disk.

Signed-off-by: Jeremiah Snapp <jeremiah@chef.io>

* Update dmg unmount logging to start at attempt 1

The logic works great as-is, but since it starts at 0 the first attempt
logged is attempt 0.

Signed-off-by: Tim Smith <tsmith@chef.io>

---------

Signed-off-by: Scott Hain <shain@chef.io>
Signed-off-by: Tim Smith <tsmith@chef.io>
Signed-off-by: Jeremiah Snapp <jeremiah@chef.io>
Signed-off-by: Jaymala Sinha <jsinha@chef.io>
Co-authored-by: Scott Hain <shain@chef.io>
Co-authored-by: Tim Smith <tsmith@chef.io>
Co-authored-by: Jeremiah Snapp <jeremiah@chef.io>
Co-authored-by: Jaymala Sinha <jsinha@chef.io>
Co-authored-by: Tim Smith <tsmith84@gmail.com>