omnibus/cli: only match override value against regex for strings
As of Ruby 3.2, `true =~ /regex/` raises an exception (a NoMethodError
for `=~`) instead of returning `nil` (seen in 3.1 and earlier),
so using `--override key:true` (and false / nil) fails with an
error. Instead of always comparing value against a regular expression,
only do it if the value didn't match a known true/false/nil value.

This appears to stem from <https://bugs.ruby-lang.org/issues/15231>.

Signed-off-by: Noel Cower <ncower@nil.dev>